refactor: ensure tests are in correct pkg
